add new view
mapView added
mainView edit

+++ MusicSalvationSDKVerson/src/com/LLC/MusicSelvation/Graphic.java
@@ -29,7 +29,7 @@ public class Graphic {
 		return bit;
 	}
 
-	static void drawPic(Canvas canvas,Bitmap bit,int mid_x,int mid_y,int rot,int alpha,Paint paint){
+	static void drawPic(Canvas canvas,Bitmap bit,int mid_x,int mid_y,float rot,int alpha,Paint paint){
 		paint.setAntiAlias(true);
 		paint.setAlpha(alpha);
 		float x=Coordinate.CoordinateX(mid_x),y=Coordinate.CoordinateY(mid_y);
